+++ src/modules/dateTime/directives/dateTimeInput/dateTimeInput.directive.ts
@@ -1,12 +1,12 @@
-import {Directive, ElementRef, EventEmitter, ExistingProvider, forwardRef, Inject, Input, OnDestroy} from '@angular/core';
-import {BindThis, isString} from '@jscrpt/common';
+import {Directive, ElementRef, EventEmitter, ExistingProvider, forwardRef, Inject, OnDestroy} from '@angular/core';
+import {BindThis} from '@jscrpt/common';
 
 import {DateTimeInput, FormatProvider} from '../../../../interfaces';
-import {DateTimeValueFormat} from '../../../../misc/enums';
 import {DATE_API, DATE_TIME_INPUT, FORMAT_PROVIDER} from '../../../../misc/tokens';
 import {DateTimeInputOutputValue, DateTimeObjectValue} from '../../../../misc/types';
 import {formatDateTime, parseDateTime} from '../../../../misc/utils';
 import {DateApi} from '../../../../services';
+import {DateTimeBase} from '../dateTimeBase';
 
 //TODO: range is unimplemented
 
@@ -25,25 +25,8 @@ import {DateApi} from '../../../../services';
         },
     ],
 })
-export class DateTimeInputDirective<TDate = unknown> implements DateTimeInput, OnDestroy
+export class DateTimeInputDirective<TDate = unknown> extends DateTimeBase<TDate> implements DateTimeInput, OnDestroy
 {
-    //######################### private fields #########################
-
-    /**
-     * Date time value format which is being worked with in this date time
-     */
-    private _valueFormat: DateTimeValueFormat = DateTimeValueFormat.DateInstance;
-
-    /**
-     * Value of date time
-     */
-    private _value: DateTimeInputOutputValue<TDate>|undefined|null;
-
-    /**
-     * Format of string representation of date
-     */
-    private _format: keyof FormatProvider = 'date';
-
     //######################### protected properties #########################
 
     /**
@@ -68,13 +51,13 @@ export class DateTimeInputDirective<TDate = unknown> implements DateTimeInput, O
     /**
      * @inheritdoc
      */
-    public get value(): DateTimeInputOutputValue<TDate>|undefined|null
+    public override get value(): DateTimeInputOutputValue<TDate>|undefined|null
     {
-        return this._value;
+        return this.ɵValue;
     }
-    public set value(value: DateTimeInputOutputValue<TDate>|undefined|null)
+    public override set value(value: DateTimeInputOutputValue<TDate>|undefined|null)
     {
-        this._value = value;
+        this.ɵValue = value;
 
         //accepts all available formats
         this.internalValue = parseDateTime(value, this.dateApi, null, this.customFormat);
@@ -117,53 +100,13 @@ export class DateTimeInputDirective<TDate = unknown> implements DateTimeInput, O
      */
     public blur: EventEmitter<FocusEvent> = new EventEmitter<FocusEvent>();
 
-    //######################### public properties - inputs #########################
-
-    /**
-     * Gets or sets date time value format which is being worked with in this date time
-     */
-    @Input()
-    public get valueFormat(): DateTimeValueFormat
-    {
-        return this._valueFormat;
-    }
-    public set valueFormat(value: DateTimeValueFormat)
-    {
-        if(isString(value))
-        {
-            this._valueFormat = DateTimeValueFormat[value] as unknown as DateTimeValueFormat;
-
-            return;
-        }
-
-        this._valueFormat = value;
-    }
-
-    /**
-     * Gets or sets format of string representation of date
-     */
-    @Input()
-    public get format(): keyof FormatProvider
-    {
-        return this._format;
-    }
-    public set format(value: keyof FormatProvider)
-    {
-        this._format = value;
-        this.customFormat = this.dateApi.getFormat(this.formatProvider[value]);
-    }
-
-    /**
-     * Custom format string representation of date
-     */
-    @Input()
-    public customFormat: string = this.dateApi.getFormat(this.formatProvider[this._format]);
-
     //######################### constructors #########################
     constructor(protected element: ElementRef<HTMLInputElement>,
-                @Inject(DATE_API) protected dateApi: DateApi<TDate>,
-                @Inject(FORMAT_PROVIDER) protected formatProvider: FormatProvider,)
+                @Inject(DATE_API) dateApi: DateApi<TDate>,
+                @Inject(FORMAT_PROVIDER) formatProvider: FormatProvider,)
     {
+        super(dateApi, formatProvider);
+
         this.element.nativeElement.addEventListener('input', this.handleInput);
         this.element.nativeElement.addEventListener('focus', this.handleFocus);
         this.element.nativeElement.addEventListener('blur', this.handleBlur);
@@ -190,7 +133,7 @@ export class DateTimeInputDirective<TDate = unknown> implements DateTimeInput, O
     protected handleInput(): void
     {
         this.internalValue = parseDateTime(this.rawValue, this.dateApi, this.valueFormat, this.customFormat);
-        this._value = formatDateTime(this.internalValue, this.valueFormat, this.customFormat);
+        this.ɵValue = formatDateTime(this.internalValue, this.valueFormat, this.customFormat);
 
         this.valueChange.next();
     }
@@ -214,11 +157,4 @@ export class DateTimeInputDirective<TDate = unknown> implements DateTimeInput, O
     {
         this.blur.emit(event);
     }
-
-    //######################### ng language server #########################
-    
-    /**
-     * Custom input type for `valueFormat` input
-     */
-    public static ngAcceptInputType_valueFormat: keyof typeof DateTimeValueFormat|DateTimeValueFormat;
 }